Fixed more size_t warnings
git-svn-id: http://openmesh.org/svnrepo/OpenMesh@559 fdac6126-5c0c-442c-9429-916003d36597

+++ trunk/src/OpenMesh/Core/IO/reader/PLYReader.cc
@@ -892,7 +892,7 @@ bool _PLYReader_::can_u_read(std::istream& _is) const {
         return false;
     }
 
-    unsigned int streamPos = _is.tellg();
+    size_t streamPos = _is.tellg();
     _is >> keyword;
     while (keyword != "end_header") {
